Removing duplicated and unused code.
