+++ src/GroupDocs.Annotation.MVC.csproj
@@ -146,7 +146,7 @@
     <Compile Include="App_Start\RouteConfig.cs" />
     <Compile Include="App_Start\WebApiConfig.cs" />
     <Compile Include="Controllers\AnnotationController.cs" />
-    <Compile Include="Products\Annotation\Annotator\AbstractSvgAnnotator.cs" />
+    <Compile Include="Products\Annotation\Annotator\AbstractTextAnnotator.cs" />
     <Compile Include="Products\Annotation\Annotator\AnnotatorFactory.cs" />
     <Compile Include="Products\Annotation\Annotator\AreaAnnotator.cs" />
     <Compile Include="Products\Annotation\Annotator\ArrowAnnotator.cs" />
